Remove VmController and rename it to vmManager
- Also sort imports
- Update logging for the controller_sync_status.go
- Fix lint errors

+++ pkg/controllers/cloud/virtualmachine_controller.go
@@ -23,7 +23,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/watch"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	antreav1alpha2 "antrea.io/antrea/pkg/apis/crd/v1alpha2"
@@ -38,11 +37,13 @@ const (
 	converterChannelBuffer = 50
 )
 
-// VirtualMachineReconciler reconciles a VirtualMachine object.
-type VirtualMachineReconciler struct {
+var vmManager *VirtualMachineManager
+
+// VirtualMachineManager reconciles a VirtualMachine object.
+type VirtualMachineManager struct {
 	client.Client
-	Log    logr.Logger
-	Scheme *runtime.Scheme
+	log    logr.Logger
+	scheme *runtime.Scheme
 
 	// TODO: pass pointer of Inventory.
 	Inventory inventory.Inventory
@@ -52,38 +53,43 @@ type VirtualMachineReconciler struct {
 
 // nolint:lll
 
-func (r *VirtualMachineReconciler) Reconcile(_ context.Context, req ctrl.Request) (ctrl.Result, error) {
-	// TODO: Remove Reconciler for VirtualMachine
-	_ = r.Log.WithValues("Virtualmachine", req.NamespacedName)
-	return ctrl.Result{}, nil
+func init() {
+	vmManager = &VirtualMachineManager{}
 }
 
-// SetupWithManager registers VirtualMachineReconciler with manager.
-// It also sets up converter channels to forward VM events.
-func (r *VirtualMachineReconciler) SetupWithManager(mgr ctrl.Manager) error {
+// GetVirtualMachineManagerInstance gets the vmManager instance.
+func GetVirtualMachineManagerInstance() *VirtualMachineManager {
+	return vmManager
+}
+
+// Configure configures the client, scheme, log and converter.
+func (r *VirtualMachineManager) Configure(client client.Client, scheme *runtime.Scheme,
+	inventory inventory.Inventory) *VirtualMachineManager {
+	r.Client = client
+	r.scheme = scheme
+	r.Inventory = inventory
+	r.log = logging.GetLogger("controllers").WithName("VirtualMachine")
 	r.converter = converter.VMConverter{
 		Client: r.Client,
 		Log:    logging.GetLogger("converter").WithName("VMConverter"),
 		Ch:     make(chan watch.Event, converterChannelBuffer),
-		Scheme: r.Scheme,
+		Scheme: r.scheme,
 	}
-	return mgr.Add(r)
+	return r
 }
 
 // Start performs the initialization of the controller.
 // VM controller is said to be initialized and synced when all the VM and EE
 // CRs are reconciled.
-func (r *VirtualMachineReconciler) Start(context.Context) error {
+func (r *VirtualMachineManager) Start() error {
 	if err := GetControllerSyncStatusInstance().waitForControllersToSync(
 		[]controllerType{ControllerTypeEE, ControllerTypeCES, ControllerTypeCPA}, syncTimeout); err != nil {
 		return err
 	}
-	r.Log.Info("Init done", "controller", ControllerTypeVM.String())
+	r.log.Info("Init done", "controller", ControllerTypeVM.String())
 
-	/* TODO: Handle restart case. Fetch all EE and all VMs (in a map) and walk thru EE and check if it exists
-	 * in the map. If not, issues a delete for that EE.
-	 */
-	if err := r.syncExternalEntities(); err != nil {
+	vmMap := r.getAllVMs()
+	if err := r.syncExternalEntities(vmMap); err != nil {
 		return err
 	}
 
@@ -103,7 +109,7 @@ func (r *VirtualMachineReconciler) Start(context.Context) error {
 }
 
 // getAllVMs returns a map which contains all the VMs.
-func (r *VirtualMachineReconciler) getAllVMs() map[types.NamespacedName]struct{} {
+func (r *VirtualMachineManager) getAllVMs() map[types.NamespacedName]struct{} {
 	vmObjList := r.Inventory.GetAllVms()
 	tempVmMap := map[types.NamespacedName]struct{}{}
 	for _, vmObj := range vmObjList {
@@ -119,20 +125,24 @@ func (r *VirtualMachineReconciler) getAllVMs() map[types.NamespacedName]struct{}
 
 // syncExternalEntities validates that each EE has corresponding VM. If it does not exist then
 // the EE will be deleted.
-func (r *VirtualMachineReconciler) syncExternalEntities() error {
-	vmMap := r.getAllVMs()
+func (r *VirtualMachineManager) syncExternalEntities(vmMap map[types.NamespacedName]struct{}) error {
 	eeList := &antreav1alpha2.ExternalEntityList{}
 	if err := r.Client.List(context.TODO(), eeList, &client.ListOptions{}); err != nil {
 		return err
 	}
 	for _, ee := range eeList.Items {
+		eeLabelKeyName, exists := ee.Labels[config.ExternalEntityLabelKeyName]
+		if !exists {
+			// Ignore EE objects not created by converter module.
+			continue
+		}
 		eeNn := types.NamespacedName{
-			Name:      ee.Labels[config.ExternalEntityLabelKeyName],
+			Name:      eeLabelKeyName,
 			Namespace: ee.Namespace,
 		}
 		if _, ok := vmMap[eeNn]; !ok {
-			r.Log.Info("Could not find matching VM object, deleting ExternalEntity", "namespacedName", eeNn)
-			// delete the external entity, since no matching VM found.
+			r.log.Info("Could not find matching VM object, deleting ExternalEntity", "namespacedName", eeNn)
+			// Delete the external entity, since no matching VM found.
 			_ = r.Client.Delete(context.TODO(), &ee)
 			continue
 		}
@@ -141,10 +151,10 @@ func (r *VirtualMachineReconciler) syncExternalEntities() error {
 }
 
 // resetWatcher sets a watcher to watch VM events.
-func (r *VirtualMachineReconciler) resetWatcher() (err error) {
+func (r *VirtualMachineManager) resetWatcher() (err error) {
 	r.vmWatcher, err = r.Inventory.WatchVms(context.TODO(), "", labels.Everything(), fields.Everything())
 	if err != nil {
-		r.Log.Error(err, "failed to start vmWatcher")
+		r.log.Error(err, "failed to start vmWatcher")
 		return err
 	}
 	return nil
@@ -152,23 +162,20 @@ func (r *VirtualMachineReconciler) resetWatcher() (err error) {
 
 // processEvent waits for any VM events and feeds the event to the
 // converter module.
-func (r *VirtualMachineReconciler) processEvent() error {
+func (r *VirtualMachineManager) processEvent() error {
 	resultCh := r.vmWatcher.ResultChan()
 	for {
-		select {
-		case event, ok := <-resultCh:
-			if !ok {
-				// TODO: we cannot exit here? Retry watcher
-				r.vmWatcher.Stop()
-				if err := r.resetWatcher(); err != nil {
-					return err
-				}
+		event, ok := <-resultCh
+		if !ok {
+			r.vmWatcher.Stop()
+			if err := r.resetWatcher(); err != nil {
+				return err
 			}
-			// Skip handling Bookmark events.
-			if event.Type == watch.Bookmark {
-				continue
-			}
-			r.converter.Ch <- event
 		}
+		// Skip handling Bookmark events.
+		if event.Type == watch.Bookmark {
+			continue
+		}
+		r.converter.Ch <- event
 	}
 }
